+++ src/Service/Athena/src/ValueObject/EngineConfiguration.php
@@ -11,7 +11,7 @@ final class EngineConfiguration
 {
     /**
      * The number of DPUs to use for the coordinator. A coordinator is a special executor that orchestrates processing work
-     * and manages other executors in a notebook session.
+     * and manages other executors in a notebook session. The default is 1.
      */
     private $coordinatorDpuSize;
 
@@ -22,7 +22,7 @@ final class EngineConfiguration
 
     /**
      * The default number of DPUs to use for executors. An executor is the smallest unit of compute that a notebook session
-     * can request from Athena.
+     * can request from Athena. The default is 1.
      */
     private $defaultExecutorDpuSize;