+++ src/Service/Athena/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Added
+
+- Added `us-iso-east-1` region
+
 ## 0.1.0
 
 First version